+++ pkg/distributor/distributor_test.go
@@ -8483,6 +8483,7 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 		expectedError                 error
 		verifyCleanUpFunc             func(func(error), *mockPermit)
 		expectedRejectedRequestsCount int
+		expectedAcquiredPermit        bool
 	}
 
 	testCases := map[string]testCase{
@@ -8499,6 +8500,7 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 				require.NotNil(t, cleanUp)
 			},
 			expectedRejectedRequestsCount: 0,
+			expectedAcquiredPermit:        true,
 		},
 		"reactive limiter enabled but cannot acquire permit": {
 			reactiveLimiterEnabled:        true,
@@ -8517,6 +8519,7 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 				require.False(t, permit.dropCalled)
 			},
 			expectedRejectedRequestsCount: 0,
+			expectedAcquiredPermit:        true,
 		},
 		"cleanup function calls permit.Drop() on context.Canceled error": {
 			reactiveLimiterEnabled:    true,
@@ -8529,6 +8532,7 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 				require.True(t, permit.dropCalled)
 			},
 			expectedRejectedRequestsCount: 0,
+			expectedAcquiredPermit:        true,
 		},
 		"cleanup function calls permit.Drop() on regular error with canceled context": {
 			reactiveLimiterEnabled:    true,
@@ -8542,6 +8546,7 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 				require.True(t, permit.dropCalled)
 			},
 			expectedRejectedRequestsCount: 0,
+			expectedAcquiredPermit:        true,
 		},
 		"cleanup function calls permit.Record() on regular error with non-canceled context": {
 			reactiveLimiterEnabled:    true,
@@ -8554,6 +8559,7 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 				require.False(t, permit.dropCalled)
 			},
 			expectedRejectedRequestsCount: 0,
+			expectedAcquiredPermit:        true,
 		},
 	}
 
@@ -8584,6 +8590,7 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 
 			// Create context
 			ctx := user.InjectOrgID(context.Background(), "user")
+			ctx = context.WithValue(ctx, requestStateKey, &requestState{})
 
 			// Cancel context if needed for testing
 			var cancel context.CancelFunc
@@ -8594,6 +8601,11 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 
 			// Call PreparePushRequest
 			cleanupFunc, err := ds[0].PreparePushRequest(ctx)
+
+			rs, ok := ctx.Value(requestStateKey).(*requestState)
+			require.True(t, ok)
+			require.Equal(t, tc.expectedAcquiredPermit, rs.reactiveLimiterPermitAcquired)
+			require.Nil(t, rs.reactiveLimiterCleanup)
 
 			// Check error
 			if tc.expectedError != nil {
@@ -8611,6 +8623,101 @@ func TestDistributor_PreparePushRequest(t *testing.T) {
 			finalRejectedCount := testutil.ToFloat64(ds[0].rejectedRequests.WithLabelValues(reasonDistributorMaxInflightPushRequests))
 			require.Equal(t, float64(tc.expectedRejectedRequestsCount), finalRejectedCount-initialRejectedCount, "rejected requests count should match")
 		})
+	}
+}
+
+func TestDistributor_AcquireReactiveLimiterPermitIdempotent(t *testing.T) {
+	testCases := map[string]struct {
+		enabled         bool
+		addRequestState bool
+		registerCleanup bool
+		expectedError   error
+	}{
+		"no permit when reactive limiter is disabled": {
+			enabled: false,
+		},
+		"context with no request state causes an error": {
+			enabled:         true,
+			addRequestState: false,
+			expectedError:   errMissingRequestState,
+		},
+		"happy case with register cleanup": {
+			enabled:         true,
+			addRequestState: true,
+			registerCleanup: true,
+		},
+		"happy case without register cleanup": {
+			enabled:         true,
+			addRequestState: true,
+			registerCleanup: false,
+		},
+	}
+
+	for testName, tc := range testCases {
+		t.Run(testName, func(t *testing.T) {
+			var limits validation.Limits
+			flagext.DefaultValues(&limits)
+
+			// Prepare distributor
+			ds, _, _, _ := prepare(t, prepConfig{
+				numDistributors: 1,
+				limits:          &limits,
+				enableTracker:   true,
+			})
+
+			// Setup reactive limiter if needed
+			if tc.enabled {
+				mockLimiter := &mockReactiveLimiter{
+					canAcquire: true,
+					permit:     &mockPermit{},
+				}
+				ds[0].reactiveLimiter = mockLimiter
+			}
+
+			// Create context
+			ctx := user.InjectOrgID(context.Background(), "user")
+			if !tc.enabled {
+				cleanup, err := ds[0].acquireReactiveLimiterPermit(ctx, tc.registerCleanup)
+				require.NoError(t, err)
+				require.Nil(t, cleanup)
+			} else if !tc.addRequestState {
+				cleanup, err := ds[0].acquireReactiveLimiterPermit(ctx, tc.registerCleanup)
+				require.Error(t, err)
+				require.ErrorIs(t, errMissingRequestState, err)
+				require.Nil(t, cleanup)
+			} else {
+				ctx = context.WithValue(ctx, requestStateKey, &requestState{})
+				checkRequestState(t, ctx, false, false)
+
+				// First call to acquireReactiveLimiterPermit should actually get a new permit.
+				cleanup, err := ds[0].acquireReactiveLimiterPermit(ctx, tc.registerCleanup)
+				require.NoError(t, err)
+				require.NotNil(t, cleanup)
+				checkRequestState(t, ctx, true, tc.registerCleanup)
+
+				// Second call to acquireReactiveLimiterPermit should fail.
+				cleanup, err = ds[0].acquireReactiveLimiterPermit(ctx, false)
+				require.Error(t, err)
+				require.ErrorIs(t, err, errReactiveLimiterPermitAlreadyAcquired)
+				require.Nil(t, cleanup)
+				checkRequestState(t, ctx, true, tc.registerCleanup)
+			}
+		})
+	}
+}
+
+func checkRequestState(t *testing.T, ctx context.Context, acquiredPermit bool, registeredCleanup bool) {
+	rs, ok := ctx.Value(requestStateKey).(*requestState)
+	require.True(t, ok)
+	if !acquiredPermit {
+		require.False(t, rs.reactiveLimiterPermitAcquired)
+		return
+	}
+	require.True(t, rs.reactiveLimiterPermitAcquired)
+	if registeredCleanup {
+		require.NotNil(t, rs.reactiveLimiterCleanup)
+	} else {
+		require.Nil(t, rs.reactiveLimiterCleanup)
 	}
 }